requirements.txt
SpeechRecognition
pyttsx3
python-dotenv
requests
gTTS
pygame
openai
pyaudio